python code restructured
