+++ app/blueprints/__init__.py
@@ -1,0 +1,3 @@
+from .charedit import *
+from .auth  import *
+from .api import *
