+++ app/models/__init__.py
@@ -1,0 +1,5 @@
+
+from .user import *
+from .character import *
+from .party import *
+
